backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
starlette==0.41.2
python-multipart==0.0.12
supabase==2.9.1
PyJWT[crypto]==2.9.0
cryptography==43.0.3
svix==1.24.0
redis[asyncio]==5.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
httpx==0.27.2
python-dotenv==1.0.1
anyio==4.6.2
